Change in score percentage

+++ api/Models/Entities/Score.cs
@@ -4,7 +4,7 @@ public class Score
 {
     public int Id { get; set; }
     public DateTime DateOfCompletion { get; set; }
-    public int PercentOfCorrectAnswers { get; set; }
+    public decimal PercentOfCorrectAnswers { get; set; }
     public virtual Quiz Quiz { get; set; }
     public int QuizId { get; set; }
     public virtual User User { get; set; }
